mds/journal: extract event serializers
